Found restaurant doesn't contain reservations data field
